# Implement the container_clone using CLONE_PARENT