--- a/crates/libcontainer/Cargo.toml
+++ b/crates/libcontainer/Cargo.toml
@@ -48,6 +48,7 @@ wasmer-wasi = { version = "2.3.0", optional = true }
 wasmedge-sdk = { version = "0.7.1", optional = true }
 wasmtime = {version = "6.0.0", optional = true }
 wasmtime-wasi = {version = "6.0.0", optional = true }
+clone3 = "0.2.3"
 
 [dev-dependencies]
 oci-spec = { version = "^0.6.0", features = ["proptests", "runtime"] }