This is the first step into fixing #1601 which implements a `container_clone` call that uses clone and `CLONE_PARENT`. Since this function uses `CLONE_PARENT` which creates sibling process instead of child process, we need to re-examine the overall logic to make sure all the `waitpid` calls are still correct, since a process can only wait on its immediate child process.