@utam0k 2023-03-01
@yihuaf How about using `clone3(2)`?