@yihuaf 2023-03-01
> @yihuaf How about using `clone3(2)`? https://github.com/utam0k/youki/blob/d1def1119b159c6916ed5ffb0da31da73c918c80/crates/libcontainer/src/process/fork.rs#L34-L63

Good point. Let me take a look tomorrow.