@yihuaf 2023-03-02
@utam0k I looked into `clone3` and I am a little hesitant to use it for this specific usecase. First, `clone3` doesn't add anything extra compared to `clone` syscall for this usecase. It is a nicer interface, but using both `clone` or `clone3` can get the job done. Second, the crate used for the `clone3` is not merged into the `libc` crate, which raises some concern. If we decide to use `clone3`, I would recommend vendoring the crate into our code base to be maintained. Lastly, should we fall back to `clone` in the case where `clone3` fails or not supported? If so, then I would argue that it will be easier to just use clone for all cases. `clone` being older, is definitely more mature and more well documented.

I don't have a strong preference. If you don't have a strong preference, I would recommend us to stay with `clone`. Given the abstraction we have built around `fork.rs`, we can easily make the switch at a later point. However, if you feel strongly about `clone3`, then I am happy to update this PR. Your call :)