@yihuaf 2023-03-02
Plus, I am hitting this with specifying both CLONE_PARENT and exit signal (SIGCHLD) with `clone3`. https://www.spinics.net/lists/linux-man/msg23564.html