@yihuaf 2023-03-02
Aha, following this thread: https://github.com/checkpoint-restore/criu/issues/926

It looks like we should not pass in exit signal when CLONE_PARENT is used. There is more context in the referenced thread.