@yihuaf 2023-03-03
> I am in favor of using clone3. We have decided that we will support kernels >= 5.4 only and clone3 has been available since kernel 5.3 so we should not need a fallback.

Sounds good. I updated the comments with what has been discussed above. PTAL. Thanks.