@utam0k 2023-03-04
> I am in favor of using clone3. We have decided that we will support kernels >= 5.4 only and clone3 has been available since kernel 5.3 so we should not need a fallback.

:100: Since they are too new, there are some options we have to be careful to use.