@yihuaf 2023-03-06
> hyperfine --prepare 'sudo sync; echo 3 | sudo tee /proc/sys/vm/drop_caches' --warmup 10 --min-runs 100 'sudo ./youki create -b tutorial a && sudo ./youki start a && sudo ./youki delete -f a'

With this PR.
```
Benchmark 1: sudo ./youki create -b tutorial a && sudo ./youki start a && sudo ./youki delete -f a
  Time (mean ± σ):     233.5 ms ±  13.9 ms    [User: 7.5 ms, System: 8.0 ms]
  Range (min … max):   197.7 ms … 265.5 ms    100 runs
```

Without the PR:
```
Benchmark 1: sudo ./youki create -b tutorial a && sudo ./youki start a && sudo ./youki delete -f a
  Time (mean ± σ):     231.5 ms ±  14.5 ms    [User: 7.4 ms, System: 7.9 ms]
  Range (min … max):   198.0 ms … 261.5 ms    100 runs
```

I believe the result matches the theoretical assumption that the performance different should be minimal since the underlying syscall is the same. The average with clone3 seems to be a little slower (about 1 ms) difference, but this should be within the standard deviation.